+++ CHANGELOG.md
@@ -28,6 +28,7 @@ and this project adheres to
 - Add ieee/;
 - Add init/kernel/cmdline.md;
 #### Changed
+- oom: add adaptive-oom-score sub-dir;
 - kernel.mk: Rename kernel version compare functions.
 - cuda/samples: all apply `cuda_compat.h`;
 - fakeCUDA: Move compat.hpp to cuda/`cuda_compat.h`;